printing the figures as jpegs
Check that this works. I can't run it here, but I think it should. As far as I can tell, savefig doesn't support jpeg, so I saved it originally as a png and then converted to a jpeg. It should save a figure with the same filename as the chart title......*should*....